+++ apps/server/src/modules/oauth/uc/oauth.uc.ts
@@ -14,7 +14,7 @@ import { SystemDto } from '@src/modules/system/service/dto/system.dto';
 import { UserService } from '@src/modules/user';
 import { UserMigrationService } from '@src/modules/user-login-migration';
 import { SchoolMigrationService } from '@src/modules/user-login-migration/service';
-import { MigrationDto } from '@src/modules/user-login-migration/service/dto/migration.dto';
+import { MigrationDto } from '@src/modules/user-login-migration/service/dto';
 import { nanoid } from 'nanoid';
 import { AuthorizationParams } from '../controller/dto';
 import { OAuthTokenDto } from '../interface';
