+++ apps/server/src/modules/user-login-migration/service/dto/index.ts
@@ -1,3 +1,4 @@
 export * from './migration.dto';
 export * from './page-content.dto';
 export * from './school-migration-flags';
+export * from './oauth-migration.dto';
