+++ apps/server/src/modules/user-login-migration/service/user-login-migration.service.spec.ts
@@ -1,8 +1,10 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
+import { Configuration } from '@hpi-schul-cloud/commons/lib';
 import { ObjectId } from '@mikro-orm/mongodb';
 import { InternalServerErrorException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
-import { UserLoginMigrationDO } from '@shared/domain';
+import { EntityNotFoundError } from '@shared/common';
+import { SchoolFeatures, UserLoginMigrationDO } from '@shared/domain';
 import { Page } from '@shared/domain/domainobject/page';
 import { SchoolDO } from '@shared/domain/domainobject/school.do';
 import { UserDO } from '@shared/domain/domainobject/user.do';
@@ -12,6 +14,8 @@ import { SchoolService } from '@src/modules/school';
 import { SystemService } from '@src/modules/system';
 import { SystemDto } from '@src/modules/system/service';
 import { UserService } from '@src/modules/user';
+import { MigrationResponse } from '../../school/controller/dto';
+import { OauthMigrationDto } from './dto';
 import { UserLoginMigrationService } from './user-login-migration.service';
 
 describe('UserLoginMigrationService', () => {
@@ -263,6 +267,255 @@ describe('UserLoginMigrationService', () => {
 				const func = async () => service.findUserLoginMigrations({ userId }, {});
 
 				await expect(func).rejects.toThrow(new InternalServerErrorException('Cannot find Sanis system information.'));
+			});
+		});
+	});
+
+	const setupMigration = (possible?: boolean, mandatory?: boolean, finished?: boolean) => {
+		const testId = 'migration';
+		const testDO = new SchoolDO({
+			id: testId,
+			name: 'testDO',
+			oauthMigrationPossible: possible ? new Date() : undefined,
+			oauthMigrationMandatory: mandatory ? new Date() : undefined,
+			oauthMigrationFinished: finished ? new Date() : undefined,
+			officialSchoolNumber: '1337',
+		});
+
+		if (testDO.oauthMigrationFinished) {
+			testDO.oauthMigrationFinalFinish = new Date(
+				testDO.oauthMigrationFinished.getTime() + (Configuration.get('MIGRATION_END_GRACE_PERIOD_MS') as number)
+			);
+		}
+
+		schoolRepo.findById.mockResolvedValue(testDO);
+		schoolRepo.save.mockResolvedValue(testDO);
+
+		return { testId, testDO };
+	};
+
+	describe('setMigration is called', () => {
+		describe('when migrationflags are truthy', () => {
+			it('should set the migrationflags', async () => {
+				const { testId, testDO } = setupMigration(true, true, true);
+
+				const resp: MigrationResponse = await schoolService.setMigration(testId, true, true, true);
+
+				expect(resp).toEqual<MigrationResponse>({
+					oauthMigrationPossible: testDO.oauthMigrationPossible,
+					oauthMigrationMandatory: testDO.oauthMigrationMandatory,
+					oauthMigrationFinished: testDO.oauthMigrationFinished,
+					oauthMigrationFinalFinish: testDO.oauthMigrationFinalFinish,
+					enableMigrationStart: true,
+				});
+			});
+
+			it('should call findById with the given id', async () => {
+				const { testId } = setupMigration(true, true, true);
+
+				await schoolService.setMigration(testId, true, true, true);
+
+				expect(schoolRepo.findById).toHaveBeenCalledWith(testId);
+			});
+
+			it('should save the DO', async () => {
+				const { testId, testDO } = setupMigration(true, true, true);
+
+				await schoolService.setMigration(testId, true, true, true);
+
+				expect(schoolRepo.save).toHaveBeenCalledWith(testDO);
+			});
+
+			describe('when the school has a feature', () => {
+				it('should add the OAUTH_PROVISIONING_ENABLED feature to the schools feature list', async () => {
+					const { testId, testDO } = setupMigration(true);
+					const existingFeature: SchoolFeatures = 'otherFeature' as SchoolFeatures;
+					testDO.features = [existingFeature];
+
+					await schoolService.setMigration(testId, true, undefined, undefined);
+
+					expect(schoolRepo.save).toHaveBeenCalledWith(
+						expect.objectContaining<Partial<SchoolDO>>({
+							features: [existingFeature, SchoolFeatures.OAUTH_PROVISIONING_ENABLED],
+						})
+					);
+				});
+			});
+
+			describe('when the school has no features yet', () => {
+				it('should set the OAUTH_PROVISIONING_ENABLED feature for the school', async () => {
+					const { testId } = setupMigration(true);
+
+					await schoolService.setMigration(testId, true, undefined, undefined);
+
+					expect(schoolRepo.save).toHaveBeenCalledWith(
+						expect.objectContaining<Partial<SchoolDO>>({
+							features: [SchoolFeatures.OAUTH_PROVISIONING_ENABLED],
+						})
+					);
+				});
+			});
+		});
+
+		describe('when oauthMigrationPossible is undefined', () => {
+			it('should set oauthMigrationPossible to undefined', async () => {
+				const { testId, testDO } = setupMigration(undefined, true, true);
+
+				const resp: MigrationResponse = await schoolService.setMigration(testId, undefined, true, true);
+
+				expect(resp).toEqual<MigrationResponse>({
+					oauthMigrationPossible: testDO.oauthMigrationPossible,
+					oauthMigrationMandatory: testDO.oauthMigrationMandatory,
+					oauthMigrationFinished: testDO.oauthMigrationFinished,
+					oauthMigrationFinalFinish: testDO.oauthMigrationFinalFinish,
+					enableMigrationStart: true,
+				});
+			});
+		});
+
+		describe('when oauthMigrationMandatory is undefined', () => {
+			it('should set oauthMigrationMandatory to undefined', async () => {
+				const { testId, testDO } = setupMigration(true, undefined, true);
+
+				const resp: MigrationResponse = await schoolService.setMigration(testId, true, undefined, true);
+
+				expect(resp).toEqual<MigrationResponse>({
+					oauthMigrationPossible: testDO.oauthMigrationPossible,
+					oauthMigrationMandatory: testDO.oauthMigrationMandatory,
+					oauthMigrationFinished: testDO.oauthMigrationFinished,
+					oauthMigrationFinalFinish: testDO.oauthMigrationFinalFinish,
+					enableMigrationStart: true,
+				});
+			});
+		});
+
+		describe('when oauthMigrationFinished is undefined', () => {
+			it('should set oauthMigrationFinished to undefined', async () => {
+				const { testId, testDO } = setupMigration(true, true, undefined);
+
+				const resp: MigrationResponse = await schoolService.setMigration(testId, true, true, undefined);
+
+				expect(resp).toEqual<MigrationResponse>({
+					oauthMigrationPossible: testDO.oauthMigrationPossible,
+					oauthMigrationMandatory: testDO.oauthMigrationMandatory,
+					oauthMigrationFinished: testDO.oauthMigrationFinished,
+					oauthMigrationFinalFinish: testDO.oauthMigrationFinalFinish,
+					enableMigrationStart: true,
+				});
+			});
+		});
+
+		describe('when DO does not have oauthMigrationPossible and oauthMigrationFinished', () => {
+			it('should set oauthMigrationStart', async () => {
+				const { testId, testDO } = setupMigration(undefined, undefined, undefined);
+
+				const resp: MigrationResponse = await schoolService.setMigration(testId, true, true, undefined);
+
+				expect(resp.oauthMigrationPossible).toEqual(testDO.oauthMigrationPossible);
+				expect(resp.oauthMigrationMandatory).toEqual(testDO.oauthMigrationMandatory);
+				expect(resp.oauthMigrationFinished).toEqual(testDO.oauthMigrationFinished);
+				expect(resp.oauthMigrationFinalFinish).toEqual(testDO.oauthMigrationFinalFinish);
+				expect(testDO.oauthMigrationStart).toBeTruthy();
+				expect(resp.enableMigrationStart).toBeTruthy();
+			});
+		});
+
+		describe('when oauthMigrationFinished is false', () => {
+			it('should set oauthMigrationStart', async () => {
+				const { testId, testDO } = setupMigration(true, true, true);
+
+				const resp: MigrationResponse = await schoolService.setMigration(testId, true, true, false);
+
+				expect(resp.oauthMigrationPossible).toEqual(testDO.oauthMigrationPossible);
+				expect(resp.oauthMigrationMandatory).toEqual(testDO.oauthMigrationMandatory);
+				expect(resp.oauthMigrationFinished).toEqual(testDO.oauthMigrationFinished);
+				expect(resp.oauthMigrationFinalFinish).toEqual(undefined);
+			});
+		});
+
+		describe('when migrationflags are falsly', () => {
+			it('should not set the migrationflags', async () => {
+				const { testId, testDO } = setupMigration(false, false, false);
+
+				const resp: MigrationResponse = await schoolService.setMigration(testId, false, false, false);
+
+				expect(resp).toEqual<MigrationResponse>({
+					oauthMigrationPossible: testDO.oauthMigrationPossible,
+					oauthMigrationMandatory: testDO.oauthMigrationMandatory,
+					oauthMigrationFinished: testDO.oauthMigrationFinished,
+					oauthMigrationFinalFinish: testDO.oauthMigrationFinalFinish,
+					enableMigrationStart: true,
+				});
+			});
+
+			it('should call findById with the given id', async () => {
+				const { testId } = setupMigration(false, false, false);
+
+				await schoolService.setMigration(testId, false, false, false);
+
+				expect(schoolRepo.findById).toHaveBeenCalledWith(testId);
+			});
+
+			it('should save the DO', async () => {
+				const { testId, testDO } = setupMigration(false, false, false);
+
+				await schoolService.setMigration(testId, false, false, false);
+
+				expect(schoolRepo.save).toHaveBeenCalledWith(testDO);
+			});
+		});
+	});
+
+	describe('getMigration is called', () => {
+		let testId: string;
+		let testDO: SchoolDO;
+		beforeEach(() => {
+			testId = 'migration';
+			testDO = new SchoolDO({
+				id: testId,
+				name: 'testDO',
+				oauthMigrationPossible: new Date(),
+				oauthMigrationMandatory: new Date(),
+				oauthMigrationFinished: new Date(),
+				officialSchoolNumber: '1337',
+			});
+			schoolRepo.findById.mockResolvedValue(testDO);
+		});
+
+		describe('when migrationflags and officialSchoolNumber are set and schoolId is given', () => {
+			it('should get the migrationflags', async () => {
+				const resp: OauthMigrationDto = await schoolService.getMigration(testId);
+
+				expect(resp.oauthMigrationPossible).toEqual(testDO.oauthMigrationPossible);
+				expect(resp.oauthMigrationMandatory).toEqual(testDO.oauthMigrationMandatory);
+				expect(resp.oauthMigrationFinished).toEqual(testDO.oauthMigrationFinished);
+				expect(resp.enableMigrationStart).toBeTruthy();
+			});
+		});
+
+		describe('when migrationflags and officialSchoolNumber are not set and schoolId is given', () => {
+			it('should get the migrationflags when not set', async () => {
+				testDO.oauthMigrationPossible = undefined;
+				testDO.oauthMigrationMandatory = undefined;
+				testDO.oauthMigrationFinished = undefined;
+				testDO.officialSchoolNumber = undefined;
+
+				const resp: OauthMigrationDto = await schoolService.getMigration(testId);
+
+				expect(resp.oauthMigrationPossible).toBeUndefined();
+				expect(resp.oauthMigrationMandatory).toBeUndefined();
+				expect(resp.oauthMigrationFinished).toBeUndefined();
+				expect(resp.enableMigrationStart).toBeFalsy();
+			});
+		});
+
+		describe('when no schoolDO is found', () => {
+			it('should throw an Exception', async () => {
+				schoolRepo.findById.mockRejectedValue(new EntityNotFoundError('School'));
+
+				const resp: Promise<OauthMigrationDto> = schoolService.getMigration('undefined');
+
+				await expect(resp).rejects.toThrow(EntityNotFoundError);
 			});
 		});
 	});
