+++ apps/server/src/shared/domain/domainobject/school.do.ts
@@ -47,8 +47,6 @@ export class SchoolDO extends BaseDO {
 
 	schoolYear?: SchoolYear;
 
-	userLoginMigrationId?: EntityId;
-
 	constructor(params: SchoolDO) {
 		super();
 		this.id = params.id;
@@ -66,6 +64,5 @@ export class SchoolDO extends BaseDO {
 		this.officialSchoolNumber = params.officialSchoolNumber;
 		this.schoolYear = params.schoolYear;
 		this.systems = params.systems;
-		this.userLoginMigrationId = params.userLoginMigrationId;
 	}
 }
