+++ apps/server/src/shared/domain/entity/all-entities.ts
@@ -2,10 +2,12 @@ import { ShareToken } from '@src/modules/sharing/entity/share-token.entity';
 import { Account } from './account.entity';
 import { Board } from './board.entity';
 import { BoardElement, LessonBoardElement, TaskBoardElement } from './boardelement.entity';
+import { BoardNode, CardNode, ColumnBoardNode, ColumnNode, TextElementNode } from './boardnode';
+import { CardElement, RichTextCardElement } from './card-element.entity';
 import { Course } from './course.entity';
-import { CourseExternalTool, ExternalTool, SchoolExternalTool } from './external-tools';
 import { CourseGroup } from './coursegroup.entity';
 import { DashboardGridElementModel, DashboardModelEntity } from './dashboard.model.entity';
+import { CourseExternalTool, ExternalTool, SchoolExternalTool } from './external-tools';
 import { File } from './file.entity';
 import { ImportUser } from './import-user.entity';
 import { Lesson } from './lesson.entity';
@@ -19,14 +21,12 @@ import { SchoolYear } from './schoolyear.entity';
 import { StorageProvider } from './storageprovider.entity';
 import { Submission } from './submission.entity';
 import { System } from './system.entity';
-import { Task } from './task.entity';
 import { TaskCard } from './task-card.entity';
+import { Task } from './task.entity';
 import { Team, TeamUser } from './team.entity';
+import { UserLoginMigration } from './user-login-migration.entity';
 import { User } from './user.entity';
 import { VideoConference } from './video-conference.entity';
-import { CardElement, RichTextCardElement } from './card-element.entity';
-import { BoardNode } from './boardnode/boardnode.entity';
-import { CardNode, ColumnBoardNode, ColumnNode, TextElementNode } from './boardnode';
 
 export const ALL_ENTITIES = [
 	Account,
@@ -72,5 +72,6 @@ export const ALL_ENTITIES = [
 	TeamNews,
 	TeamUser,
 	User,
+	UserLoginMigration,
 	VideoConference,
 ];
