+++ apps/server/src/shared/repo/school/school.repo.ts
@@ -3,7 +3,6 @@ import { EntityManager } from '@mikro-orm/mongodb';
 import { Injectable, InternalServerErrorException } from '@nestjs/common';
 import { EntityId, ISchoolProperties, School, System } from '@shared/domain';
 import { SchoolDO } from '@shared/domain/domainobject/school.do';
-import { UserLoginMigration } from '@shared/domain/entity/user-login-migration.entity';
 import { Logger } from '@src/core/logger';
 import { BaseDORepo } from '../base.do.repo';
 
@@ -50,7 +49,6 @@ export class SchoolRepo extends BaseDORepo<SchoolDO, School, ISchoolProperties> 
 			officialSchoolNumber: entity.officialSchoolNumber,
 			schoolYear: entity.schoolYear,
 			systems: entity.systems.isInitialized() ? entity.systems.getItems().map((system: System) => system.id) : [],
-			userLoginMigrationId: entity.userLoginMigration?.id,
 		});
 	}
 
@@ -67,9 +65,6 @@ export class SchoolRepo extends BaseDORepo<SchoolDO, School, ISchoolProperties> 
 			systems: entityDO.systems
 				? entityDO.systems.map((systemId: EntityId) => this._em.getReference(System, systemId))
 				: [],
-			userLoginMigration: entityDO.userLoginMigrationId
-				? this._em.getReference(UserLoginMigration, entityDO.userLoginMigrationId)
-				: undefined,
 		};
 	}
 }
