+++ apps/server/src/shared/repo/userloginmigration/user-login-migration.repo.integration.spec.ts
@@ -2,9 +2,11 @@ import { createMock } from '@golevelup/ts-jest';
 import { EntityManager } from '@mikro-orm/mongodb';
 import { Test, TestingModule } from '@nestjs/testing';
 import { School, System, UserLoginMigrationDO } from '@shared/domain';
+import { UserLoginMigration } from '@shared/domain/entity/user-login-migration.entity';
 import { MongoMemoryDatabaseModule } from '@shared/infra/database';
 import { cleanupCollections, schoolFactory, systemFactory } from '@shared/testing';
 import { Logger } from '@src/core/logger';
+import { userLoginMigrationFactory } from '../../testing/factory/user-login-migration.factory';
 import { UserLoginMigrationRepo } from './user-login-migration.repo';
 
 describe('UserLoginMigrationRepo', () => {
@@ -54,6 +56,7 @@ describe('UserLoginMigrationRepo', () => {
 				});
 
 				await em.persistAndFlush([school, sourceSystem, targetSystem]);
+				em.clear();
 
 				return {
 					domainObject,
@@ -68,6 +71,80 @@ describe('UserLoginMigrationRepo', () => {
 
 				expect(result).toMatchObject(expected);
 				expect(result.id).toBeDefined();
+			});
+		});
+	});
+
+	describe('delete', () => {
+		describe('when saving a UserLoginMigrationDO', () => {
+			const setup = async () => {
+				const userLoginMigration: UserLoginMigration = userLoginMigrationFactory.buildWithId();
+
+				await em.persistAndFlush(userLoginMigration);
+				em.clear();
+
+				const domainObject: UserLoginMigrationDO = repo.mapEntityToDO(userLoginMigration);
+
+				return {
+					userLoginMigration,
+					domainObject,
+				};
+			};
+
+			it('should not delete the school', async () => {
+				const { domainObject, userLoginMigration } = await setup();
+				const { school } = userLoginMigration;
+
+				await repo.delete(domainObject);
+
+				await em.findOneOrFail(School, { id: school.id });
+			});
+		});
+	});
+
+	describe('findBySchoolId', () => {
+		describe('when searching for a UserLoginMigration by its school id', () => {
+			const setup = async () => {
+				const userLoginMigration: UserLoginMigration = userLoginMigrationFactory.buildWithId();
+
+				await em.persistAndFlush(userLoginMigration);
+				em.clear();
+
+				const domainObject: UserLoginMigrationDO = repo.mapEntityToDO(userLoginMigration);
+
+				return {
+					userLoginMigration,
+					domainObject,
+				};
+			};
+
+			it('should return the UserLoginMigration', async () => {
+				const { domainObject, userLoginMigration } = await setup();
+
+				const result: UserLoginMigrationDO | null = await repo.findBySchoolId(userLoginMigration.school.id);
+
+				expect(result).toEqual(domainObject);
+			});
+		});
+
+		describe('when searching for a UserLoginMigration by an unknown school id', () => {
+			const setup = async () => {
+				const school: School = schoolFactory.buildWithId({ userLoginMigration: undefined });
+
+				await em.persistAndFlush(school);
+				em.clear();
+
+				return {
+					school,
+				};
+			};
+
+			it('should return null', async () => {
+				const { school } = await setup();
+
+				const result: UserLoginMigrationDO | null = await repo.findBySchoolId(school.id);
+
+				expect(result).toBeNull();
 			});
 		});
 	});
